connecthub-backend/pyproject.toml
[project]
name = "connecthub-backend"
version = "0.1.0"
description = "FastAPI backend for ConnectHub dating app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic[email]>=2.5.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.29.0",
    "python-socketio>=5.10.0",
    "redis>=6.4.0",
]